requirements.txt
numpy>=1.21.0
opencv-python>=4.5.0
tensorflow>=2.8.0
pillow>=8.0.0
flask>=2.0.0
scikit-learn>=0.24.0
torch>=1.9.0
torchvision>=0.10.0
transformers>=4.15.0
python-dotenv>=0.19.0
gevent>=24.11.1
tk>=0.1.0
scikit-image>=0.19.0
pydicom>=2.3.0
scipy>=1.7.0